requirements.txt
Django==1.11.4
djangorestframework==3.6.4
gunicorn==19.7.1
pytz==2017.2
psycopg2==2.7.3.1
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.2
django-allauth>=0.24.1
six==1.9.0
